+++ src/functions/useFormPersistSingle.ts
@@ -1,0 +1,10 @@
+import { useFormPersistSinglePage } from '../core/useFormPersistSinglePage';
+import type { UseFormReturn, FieldValues } from 'react-hook-form';
+
+export function useFormPersistSingle<T extends FieldValues>(
+  useFormReturn: UseFormReturn<T>,
+  excludes: (keyof T)[] = [],
+) {
+  const ROOT_KEY = 'RFHP_SINGLE';
+  return useFormPersistSinglePage(useFormReturn, excludes, ROOT_KEY);
+}